refactor: 컴포넌트 폴더링 구조 개선
